Move stuff around and make tests pass
