+++ spec/units/polling_spec.rb
@@ -21,12 +21,11 @@ describe ActiveEncode::Polling do
 
   describe '#after_create' do
     let(:encode_class) { PollingEncode }
-    let(:encode) { encode_class.create("sample.mp4") }
+    subject { encode_class.create("sample.mp4") }
 
     it "enqueue PollingJob after polling wait time" do
-      # expect(ActiveEncode::PollingJob).to have_been_enqueued.with(encode.id, {offset:ActiveEncode::Polling::POLLING_WAIT_TIME })
-      expect(ActiveEncode::PollingJob).to have_been_enqueued.with(encode.id)
-      expect(ActiveEncode::PollingJob).to have_been_enqueued.at(ActiveEncode::Polling::POLLING_WAIT_TIME.from_now)
+      subject
+      expect(ActiveEncode::PollingJob).to have_been_enqueued.with(subject)
     end
   end
 end
